update auth route
